Работа с пользовательскими полями вынесена в трейты
